Switch when the 'build' step happens (#72)
Currently a build is triggered every time `npm install` is run. This can
be considered a bit unexpected, and is an unnecessary step when
installing the dependencies just to run code style tests. It also seems
to stop the npm cache working.

Instead, this switches the build to happen before starting the app, in
the same way that the build happens before running `npm run watch`.