Style guide (#81)
starting point for style guide. 

Things to check:
 * all the points really
* anything re formatting. I thought this would be a definition list but
the health guide A-Z is headings?
 
Fixes #62 
 
 > ### Style guide
 >
>We normally use the [NHS content
guide](https://service-manual.nhs.uk/content) for terms. But there are
some terms that are only used in the guidance for the kit:
 >
 > ### command line
 >
>Lower case. Only use for Windows instructions. Nortmally it will be
‘the command line’.
 > 
> #### GOV.UK Prototype Kit, GOV.UK Design System and other GOV.UK
products
 >
 > Title case.
 >
 > #### NHS, NHS.UK, NHS England
 >
> Use ‘NHS.UK’ for products that are for delivery teams and not the
public (find out more on the [NHS health writing
A-Z](https://service-manual.nhs.uk/content/a-to-z-of-nhs-health-writing#N))This
includes the NHS.UK prototype kit, NHS.UK design system, NHS.UK service
manual or NHS.UK website.
 >
 > #### NHS England laptops
> Use ‘NHS England Windows laptops’ instructions for people using
corporate laptops.
 >
 > #### NHS.UK prototype kit
 >
> ‘NHS.UK’ is in caps and ‘prototype kit’ in lower case (except where
the NHS.UK is shown as a logo when it is then presented as ‘NHS
Prototype kit’). Write out in full the first time it is used on a page,
then call ’the kit’. Do not use ‘prototype kit’ if you can.
 >
 > #### terminal
 > 
> Lower case. Only use for Mac and Linux instructions. Usually it will
be ‘the terminal’.

Things I'd like to add:
* anything about file paths - for example do we always need to spell
them out as from the top or is it OK to be a bit more loose about them?
* more about the terminal and command prompt stuff, maybe I don't have
enough yet to complete that

---------

Co-authored-by: Ed Horsford <mail@edwardhorsford.com>