Added notes to the 'switching from the GOV.UK Prototype Kit' on updating the prototype kit (#143)
Added section

> ## Updating your prototype
> 
> If you want to update your prototype to the latest version, you will
need to download the latest version of the kit and copy some files into
your prototype. Go to the [guide on updating the
kit](/how-tos/updating-the-kit) for more details.

Still deciding if this could be a part of the installing and using and
if it should be a dos/don't, but at the very least this catches
something that is dramatically different.